@@ -13,18 +13,22 @@
   },
   "devDependencies": {
     "@aws-cdk/assert": "^1.15.0",
-    "@types/jest": "^24.0.18",
+    "@types/jest": "^24.0.22",
     "aws-cdk": "^1.15.0",
     "jest": "^24.9.0",
-    "ts-jest": "^24.0.2",
-    "ts-node": "^8.1.0",
-    "typescript": "~3.6.2"
+    "ts-jest": "^24.1.0",
+    "ts-node": "^8.4.1",
+    "typescript": "~3.7.2"
   },
   "dependencies": {
-    "@aws-cdk/core": "^1.15.0",
-    "@aws-cdk/aws-route53": "^1.15.0",
     "@aws-cdk/aws-cloudfront": "^1.15.0",
-    "source-map-support": "^0.5.9"
+    "@aws-cdk/aws-iam": "^1.15.0",
+    "@aws-cdk/aws-route53": "^1.15.0",
+    "@aws-cdk/aws-route53-targets": "^1.15.0",
+    "@aws-cdk/aws-s3": "^1.15.0",
+    "@aws-cdk/core": "^1.15.0",
+    "@types/node": "^12.12.6",
+    "source-map-support": "^0.5.16"
   },
   "license": "MIT-0",
   "repository": {}
